fix (#308) : move component
